requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.3.0
babel==2.16.0
beautifulsoup4==4.13.4
bleach==6.2.0
Bottleneck==1.4.2
Brotli==1.0.9
cached-property==1.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==2.19.2
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
executing==0.8.3
fastapi==0.116.1
fastjsonschema==2.20.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.3.1
gast==0.6.0
gmpy2==2.2.1
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.33.4
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.17
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.4
notebook_shim==0.2.4
numexpr==2.11.0
numpy==1.26.4
opentelemetry-api==1.35.0
opt_einsum==3.4.0
optree==0.16.0
overrides==7.4.0
packaging==25.0
pandas==2.2.3
pandas-stubs==2.3.0.250703
pandocfilters==1.5.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pip==25.1
platformdirs==4.3.7
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.3.1
protobuf==4.25.8
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyarrow-hotfix==0.7
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.22.3
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
starlette==0.47.1
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
termcolor==3.1.0
terminado==0.17.1
tf_keras==2.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.20.4
torch==2.4.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.45.2
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
